+++ src/clj/backtype/storm/daemon/task.clj
@@ -94,7 +94,8 @@
                                stream)))))
 
 (defn mk-tasks-fn [task-data]
-  (let [executor-data (:executor-data task-data)
+  (let [task-id (:task-id task-data)
+        executor-data (:executor-data task-data)
         component-id (:component-id executor-data)
         ^WorkerTopologyContext worker-context (:worker-context executor-data)
         storm-conf (:storm-conf executor-data)
@@ -127,7 +128,7 @@
                (when (= :direct grouper)
                   ;;  TODO: this is wrong, need to check how the stream was declared
                   (throw (IllegalArgumentException. "Cannot do regular emit to direct stream")))
-               (let [comp-tasks (grouper values)]
+               (let [comp-tasks (grouper task-id values)]
                  (if (or (sequential? comp-tasks) (instance? Collection comp-tasks))
                    (.addAll out-tasks comp-tasks)
                    (.add out-tasks comp-tasks)